fix: clear commit detail release blockers
