+++ internal/git/command_context_unix_test.go
@@ -1,3 +1,5 @@
+//go:build aix || darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris
+
 package git
 
 import (
