+++ internal/ui/editor_group_test.go
@@ -9,12 +9,17 @@ import (
 func TestOpenPluginTabReplacementClosesPreviousContent(t *testing.T) {
 	group := NewEditorGroupWidget(nil, 4, true, "relative")
 	closed := false
+	notified := ""
+	group.OnContentTabClose = func(id string) { notified = id }
 	first := NewCommitDetailWidget("/repo", "ref", "abcdef0", false)
 	first.OnClose = func() { closed = true }
 	group.OpenPluginTab("commit", "Commit", first)
 	group.OpenPluginTab("commit", "Replacement", NewCommitDetailWidget("/repo", "other", "1234567", false))
 	if !closed {
 		t.Fatal("replacing a content tab did not close its previous incarnation")
+	}
+	if notified != "commit" {
+		t.Fatalf("replacement close notification = %q, want commit", notified)
 	}
 }
